hw #6

+++ src/Controller/HouseController.php
@@ -4,18 +4,19 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
-use App\Entity\House;
+use App\Dto\Create\HouseCreateDto;
+use App\Dto\Output\HouseOutputDto;
+use App\Dto\Update\HouseUpdateDto;
 use App\Repository\HouseRepository;
 use App\Service\HouseService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Exception\HttpException;
-use Symfony\Component\Routing\Attribute\Route;
-use Symfony\Component\Security\Http\Attribute\IsGranted;
+use Symfony\Component\HttpKernel\Attribute\MapRequestPayload;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
-#[Route('/api/houses')]
 class HouseController extends AbstractController
 {
     public function __construct(
@@ -25,51 +26,81 @@ class HouseController extends AbstractController
     ) {
     }
 
-    #[IsGranted('ROLE_ADMIN')]
-    #[Route('/', methods: ['GET'])]
-    public function getAllHouses(): Response
+    public function getAllHouses(): JsonResponse
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedHttpException('Admin access required.');
+        }
+
         $houses = $this->houseRepository->findAll();
 
-        return $this->json($houses, Response::HTTP_OK);
+        return $this->json(HouseOutputDto::mapArray($houses));
     }
 
-    #[IsGranted('ROLE_USER')]
-    #[Route('/available', methods: ['GET'])]
-    public function getAvailableHouses(): Response
+    public function getAvailableHouses(): JsonResponse
     {
         $houses = $this->houseRepository->findAvailable();
 
-        return $this->json($houses, Response::HTTP_OK);
+        return $this->json(HouseOutputDto::mapArray($houses));
     }
 
-    #[IsGranted('ROLE_ADMIN')]
-    #[Route('/create', methods: ['POST'])]
-    public function createHouse(Request $request): Response
+    public function createHouse(#[MapRequestPayload] HouseCreateDto $dto): JsonResponse
     {
-        $requestData = $request->toArray();
-
-        if (!isset($requestData['sleeping_places'])) {
-            throw new HttpException(Response::HTTP_BAD_REQUEST, 'Missing required param: sleeping_places!');
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedHttpException('Admin access required.');
         }
 
-        $sleepingPlaces = $requestData['sleeping_places'];
-
-        $house = $this->houseService->create($sleepingPlaces);
-
+        $house = $this->houseService->create($dto->sleepingPlaces);
         $this->em->flush();
 
-        return $this->json($house, Response::HTTP_OK);
+        return $this->json(new HouseOutputDto($house), Response::HTTP_CREATED);
     }
 
-    #[IsGranted('ROLE_ADMIN')]
-    #[Route('/{id}', methods: ['DELETE'])]
-    public function removeHouse(House $house): Response
+    public function getHouseDetail(int $id): JsonResponse
     {
-        $this->houseService->remove($house);
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedHttpException('Admin access required.');
+        }
 
+        $house = $this->houseRepository->find($id);
+        if (null === $house) {
+            throw new NotFoundHttpException('House not found.');
+        }
+
+        return $this->json(new HouseOutputDto($house));
+    }
+
+    public function removeHouse(int $id): JsonResponse
+    {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedHttpException('Admin access required.');
+        }
+
+        $house = $this->houseRepository->find($id);
+        if (null === $house) {
+            throw new NotFoundHttpException('House not found.');
+        }
+
+        $this->houseService->remove($house);
         $this->em->flush();
 
-        return $this->json([], Response::HTTP_OK);
+        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
+    }
+
+    public function updateHouse(int $id, #[MapRequestPayload] HouseUpdateDto $dto): JsonResponse
+    {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedHttpException('Admin access required.');
+        }
+
+        $house = $this->houseRepository->find($id);
+        if (null === $house) {
+            throw new NotFoundHttpException('House not found.');
+        }
+
+        $house->setSleepingPlaces($dto->sleepingPlaces);
+        $this->em->flush();
+
+        return $this->json(new HouseOutputDto($house));
     }
 }
